ai/requirements.txt
# Required Python packages for OpenAI API integration
openai>=1.0.0
python-dotenv>=1.0.0

